% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govdata360stats.R
\name{gov360DataSet}
\alias{gov360DataSet}
\title{Get the metadata of the indicators}
\usage{
gov360DataSet(dsId)
}
\arguments{
\item{dsId}{The id of the dataset}
}
\description{
Get the metadata of the indicators
}
