% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govdata360search.R
\name{gov360msearch}
\alias{gov360msearch}
\title{Get the metadata of the indicators}
\usage{
gov360msearch(nameSearch = "", sourceSearch = "")
}
\arguments{
\item{nameSearch}{search indicator by name}

\item{sourceSearch}{search indicator by source}
}
\value{
Data frame with id, name, dataset, valueType, datasetID,
defaultViz, doNotUseVix, rank, definition, units, subIndicatorType, timeFrames, byProduct, byPartner, periodicity and dateRange
}
\description{
Data frame with series code, name, description, and source for the govdata360
}
\examples{
gov360msearch()
}
