% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govdata360stats.R
\name{gov360stats}
\alias{gov360stats}
\title{Pull the data for some indicators and countries.}
\usage{
gov360stats(country = "ALL", dateRange = "2010", indicatorcode = "ALL")
}
\arguments{
\item{country}{Character string. ISO3 code of the countries. E.g COL for Colombia or 'ALL' for all the countries}

\item{dateRange}{Character string. Range of years of the indicators}

\item{indicatorcode}{codes of the indicators separated by comma. 'All' for all the indicators}
}
\value{
Data frame with  CountryCode, CountryTableName, IndicatorCode, IndicatorName, TopicName, SubTopicName, Year, Quarter,
AggregationLevel, AggregatedValue, UOM
}
\description{
Data frame with the data of the specified indicators and countries
All countries and All indicators cannot be requested
When All indicators data is requested, data can be requested for a maximum of 4 countries
When All country data is requested, data can be requested for a maximum of 10 indicators (see iadbstats.list)
}
\examples{
gov360stats(country="COL",dateRange="2010",indicatorcode="27712")
gov360stats(country="ARG,COL",dateRange="2010,2011",indicatorcode="27712")
}
