% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govdata360stats.R
\name{gov360stats.list}
\alias{gov360stats.list}
\title{Pull the data for a vector of indicators and ALL the countries.}
\usage{
gov360stats.list(dateRange = "ALL", pIndicators, pCountry = "ALL")
}
\arguments{
\item{dateRange}{Character string. Range of years of the indicators}

\item{country}{Character string. ISO3 code of the countries. E.g COL for Colombia or 'ALL' for all the countries}

\item{indicatorCodes}{vector of indicator codes c("27712","27870")}
}
\value{
Data frame with  CountryCode, CountryTableName, IndicatorCode, IndicatorName, TopicName, SubTopicName, Year, Quarter, AggregationLevel, AggregatedValue, UOM
}
\description{
Data frame with the data of the specified indicators and countries
}
\examples{
codes=c("27712","27870")
gov360stats.list(pIndicators=codes)
}
